refactor(evaluatorq-py): re-point redteam blast-radius to common/tracing (RES-899)
